TileBoundingVolumeProps => core-common
